Implement get combined employee table from DB
